+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Change request types for channels, platforms, and architectures from form-data to application/json 
 
+- Add `electron-builder` and `squirrel_darwin` updaters.
+
 ## v1.4.3
 
 ### Features